factor console logging out of run.ts
